build: Let users pass in their own CFLAGS, CPPFLAGS, LDFLAGS
There is a widely-used convention that CFLAGS, CPPFLAGS and LDFLAGS are
set by the user (or packaging system) when building software,
defaulting to empty: in Autotools terminology they are said to be
"reserved for the user". Linux distributions use this to provide
globally-used compiler flags, for example enabling enabling various
hardening features.

These are conventionally placed on the compiler command-line *after* the
build system's choice of CFLAGS and LDFLAGS, so that they can override
the build system's defaults. For example, when the Debian packaging
toolchain is told to build a package without optimization for debugging,
it includes -O0 in CFLAGS. This would not be effective in previous
versions of yquake2, which hard-code -O2.

Loosely based on a patch by Fabian Greffrath, which we have been applying
in Debian for several years.

Signed-off-by: Simon McVittie <smcv@debian.org>