build: Add an option to turn off RPATHs
When distributions like Debian build binaries that are going to use the
standard system-wide libraries in directories that are searched by
default, it's undesired to have a RPATH that makes the library search
behaviour less predictable. Continue to set the RPATH by default, for
use in portable binaries, but provide an off-switch.

Based on a patch by Fabian Greffrath, which we have been applying in
Debian for several years.

Signed-off-by: Simon McVittie <smcv@debian.org>