arm: Only use yield instruction if supported
The default on 32-bit ARM is now -march=armv6k, but it's overridable,
and must be overridden on some toolchains that have a higher baseline
and cannot compile for -march=armv6k. This includes Debian armhf
(ARMv7 with the VFPv3D16 floating-point extensions), where armv6k
appears to imply an older VFP (floating point) ABI than anything the
compiler supports.

Conditionally enable the yield instruction at compile-time if the target
CPU has it. As far as I can find out, yield was added in ARMv6K, a minor
revision of ARMv6, so it should be available in all ARMv7 or higher CPUs
(which includes anything 64-bit), plus ARMv6 CPUs *if* they are actually
ARMv6K.

Signed-off-by: Simon McVittie <smcv@debian.org>