Don't force target CPU to armv8-a on aarch64
ARMv8-A is the oldest/most-limited/least-extended version of aarch64,
analogous to the earliest AMD 64-bit processor in the x86 world.
As a result, forcing this target CPU can only disable features, and
isn't going to enable anything we might want. The yield opcode was
added in ARMv7, and all aarch64 CPUs are ARMv8-A or higher (which is
a superset of ARMv7), so no special architecture specification is
necessary to access the yield opcode.

Signed-off-by: Simon McVittie <smcv@debian.org>